+++ find-core/src/main/resources/db/migration/V0__init.sql
@@ -4,8 +4,8 @@ SET SCHEMA find;
 
 CREATE TABLE searches
 (
-  search_id INT AUTO_INCREMENT PRIMARY KEY,
-  user_id INT NOT NULL,
+  search_id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  user_id BIGINT NOT NULL,
   search_type INT NOT NULL,
   title NVARCHAR(1000) NOT NULL,
   query_text NVARCHAR(21844) NOT NULL,
@@ -21,7 +21,7 @@ CREATE INDEX ix__searches__search_type ON searches (search_type);
 
 CREATE TABLE users
 (
-  user_id INT AUTO_INCREMENT PRIMARY KEY,
+  user_id BIGINT AUTO_INCREMENT PRIMARY KEY,
   domain VARCHAR(255),
   user_store VARCHAR(255),
   uuid UUID,
@@ -33,8 +33,8 @@ ALTER TABLE searches ADD CONSTRAINT fk__searches__users FOREIGN KEY (user_id) RE
 
 CREATE TABLE search_parametric_values
 (
-  search_parametric_values_id INT AUTO_INCREMENT PRIMARY KEY,
-  search_id INT NOT NULL,
+  search_parametric_values_id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  search_id BIGINT NOT NULL,
   field NVARCHAR(21844) NOT NULL,
   value NVARCHAR(21844) NOT NULL
 );
@@ -44,8 +44,8 @@ ALTER TABLE search_parametric_values ADD CONSTRAINT fk__search_parametric_values
 
 CREATE TABLE search_indexes
 (
-  search_index_id INT AUTO_INCREMENT PRIMARY KEY,
-  search_id INT NOT NULL,
+  search_index_id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  search_id BIGINT NOT NULL,
   name NVARCHAR(21844) NOT NULL,
   domain VARCHAR(255)
 );
@@ -55,8 +55,8 @@ ALTER TABLE search_indexes ADD CONSTRAINT fk__search_indexes__searches FOREIGN K
 
 CREATE TABLE search_stored_state
 (
-  search_stored_state_id INT AUTO_INCREMENT PRIMARY KEY,
-  search_id INT NOT NULL,
+  search_stored_state_id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  search_id BIGINT NOT NULL,
   state_token VARCHAR(100) NOT NULL
 );
 
@@ -65,8 +65,8 @@ ALTER TABLE search_stored_state ADD CONSTRAINT fk__search_stored_state__searches
 
 CREATE TABLE search_related_concepts
 (
-  search_related_concepts_id INT AUTO_INCREMENT PRIMARY KEY,
-  search_id INT NOT NULL,
+  search_related_concepts_id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  search_id BIGINT NOT NULL,
   concept NVARCHAR(21844) NOT NULL
 );
